requirements.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
Deprecated==1.2.14
filelock==3.12.2
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
gdown==4.7.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
joblib==1.2.0
keras==2.13.1rc0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mtcnn==0.1.1
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pillow-heif==0.11.1
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
tensorflow-aarch64==2.12.0
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
